Moved code for increasing score to the animal class, updated scoring to happen after an animal has reached full satiation, and setup score increase to be relative to the hunger difficulty of the animal.

+++ Assets/Scripts/Game Scene/UIHungerBar.cs
@@ -16,6 +16,6 @@ public class UIHungerBar : MonoBehaviour
 
     private void Update()
     {
-        slider.value = animal.Satiation / animal.Hunger;
+        slider.value = (float)animal.Satiation / (float)animal.Hunger;
     }
 }
